requirements.txt
pandas==1.0.1
pushbullet.py==0.11.0
xalpha==0.3.1
